requirements.txt
aiohappyeyeballs==2.6.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
inspect_mate==0.0.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.44
langchain-mcp-adapters==0.0.4
langchain-openai==0.3.8
langgraph==0.3.10
langgraph-checkpoint==2.0.19
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.13
markdown-it-py==3.0.0
mcp==1.3.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
openai==1.66.3
orjson==3.10.15
packaging==24.2
propcache==0.3.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
zstandard==0.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyenv==0.4.11
anyio==4.8.0
appdirs==1.4.4
argcomplete==3.6.0
attrs==25.3.0
babel==2.17.0
Beaker==1.12.1
beautifulsoup4==4.13.3
blinker==1.9.0
blivet==3.11.0
blivet-gui==2.6.0
Brlapi==0.8.5
build==1.2.1
CacheControl==0.14.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-plugins==1.1.1
configobj==5.0.8
crashtest==0.4.1
crypt_r==3.13.1
cryptography==43.0.0
cupshelpers==1.0
dasbus==1.7
dataclasses-json==0.6.7
dbus-python==1.3.2
distlib==0.3.8
distro==1.9.0
dnf==4.23.0
docopt==0.6.2
dulwich==0.21.7
fastjsonschema==2.19.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.15.4
fros==1.1
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
hishel==0.1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanize==4.9.0
idna==3.7
installer==0.7.0
iso639==0.1.4
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-mcp-adapters==0.0.5
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langgraph==0.3.11
langgraph-checkpoint==2.0.20
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.15
langtable==0.0.68
lark==1.1.9
libcomps==0.1.21
libdnf==0.74.0
libvirt-python==10.6.0
louis==3.28.0
lxml==5.2.1
Mako==1.2.3
MarkupSafe==2.1.5
marshmallow==3.26.1
mcp==1.4.1
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.30.0
networkx==3.4.2
nftables==0.1
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
openai==1.66.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
pkginfo==1.12.0
platformdirs==4.3.7
ply==3.11
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.6.0
productmd==1.43
progressbar2==3.53.2
propcache==0.3.0
protobuf==5.29.3
protonvpn_cli==2.2.12
psutil==5.9.8
ptyprocess==0.7.0
pwquality==1.4.5
pyarrow==19.0.1
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pyenchant==3.2.2
pygit2==1.17.0
Pygments==2.19.1
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.58
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyparted==3.13.0
pyproject_hooks==1.0.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-utils==3.7.0
pythondialog==3.5.3
pytz==2025.1
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
rapidfuzz==3.5.2
redis==5.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rpm==4.20.1
rpmautospec==0.7.3
rpmautospec-core==0.1.5
rstapi==1.0.4
safetensors==0.5.3
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
shellingham==1.5.4
shtab==1.7.1
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sos==4.8.2
soupsieve==2.6
SQLAlchemy==2.0.39
sse-starlette==2.2.1
starlette==0.46.1
streamlit==1.43.2
sympy==1.13.1
systemd-python==235
Tempita==0.5.2
tenacity==9.0.0
terminator==2.1.4
tiktoken==0.9.0
tokenizers==0.21.1
tokonomics==0.3.9
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.4
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
trove-classifiers==2025.3.13.13
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.6.6
uvicorn==0.34.0
virtualenv==20.21.1
watchdog==6.0.0
xkbregistry==0.3
yarl==1.18.3
zstandard==0.23.0
